requirements.txt
streamlit
chromadb
python-dotenv
google-generativeai
pypdf
tqdm
pysqlite3-binary